Move check_number_of_args to util.py
